按钮组的高亮增加对就工程的容错

+++ simulator/src/components/Simulator.js
@@ -1651,7 +1651,7 @@ module.exports =   React.createClass({
             var maxValue = widget.info.maxValue;
             // var curArc = widget.info.value;
             var curDashboardTag = this.findTagByName(widget.tag);
-            var curDashboardTagValue=parseInt((curDashboardTag&&curDashboardTag.value)||0);
+            var curDashboardTagValue=parseFloat((curDashboardTag&&curDashboardTag.value)||0);
             var curArc = (maxArc-minArc)/(maxValue-minValue)*curDashboardTagValue;
             var currentValue = curDashboardTag&&curDashboardTag.value||0;
             var clockwise = widget.info.clockwise == '1'?1:-1;
